requirements.txt
gradio>=4.0.0
openai>=1.0.0
openai-agents>=0.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
pytz>=2023.3
